Add /graph:noBuild (#6016)
This new option constructs the graph but does not build the nodes.

This is useful for a few use cases:

determine if a large repo can evaluate all of its projects (e.g., is VS setup right, are all imports magically pointing to where they should, etc)
easily investigate evaluation perf. With noBuild, the only thing that MSBuild does is evaluate all projects under a single process, so it's easier to throw it under a profiler.
in the future when project caching is available, it can be used to warm up / download the caches but not do the build.
generally makes life easier for people that only want to investigate evaluation and not the build. The binlog also shows up nicely containing all the evaluation nodes.