Remove unused files from the repo (#6158)
* Remove unused 'Tasks.UnitTests2' project

This is a part of XMake project in VS repo.
This is not used here in this MSBuild repo.

* Remove unused 'netci.groovy' file

Replaced with Azure Pipelines and Arcade Sdk

* Remove unused 'RunApiPort.bat' file

No longer needed. Replaced with dotnet tool.

* Remove unused 'runApiCompat.bat' file

No longer needed. Replaced with dotnet tool.

* Remove unused 'DefaultTasks.bat' file

This batch file was used to copy targets for VS insertion in the old MSBuild project in VS repo.
This is no longer needed as we achieve this using custom build logic in the project file itself.

* Remove unused 'CompatBaseline.txt' file

Used when merging xplat branch.
No longer needed.