Add ETW start and stop events for file changes (#16531)
* Add ETW start and stop events for file changes

* Change event structure to one start/stop event pair

This changes the event structure to be one start/stop pair that we use
in several places. We emit events for start and stop of the
TryHandleFileChange in HotReload.cs that bracket the overall time spent
handling the change. We also now emit events inside the different file
handler classes, so that we can better understand how much time was
spent in each phase.

* Change events to also fire on fail-fast

* Fix typo

* Remove syntax error

* Add copyright info and make static readonly

* Make enum arg name more descriptive