Cache resource strings in RAR logging.
Instead of retrieving resource strings every time, cache them, reduce allocations and concatenations. No need to use a resource string to just add 4/8/10/12 spaces.

Primarily this avoids concatenating the final string into a single string, but instead keeps the message separate from the arguments until the final string is requested.

Don't log the image runtime version if it's v4.0.30319 as it's the overwhelming default and carries no useful information.

Use better terminology for inclusion and exclusion lists.

This change looks extensive, but is pretty mechanical. Hard to review, but a significant improvement to allocations and RAR logging overhead.