Allow users that have certain special characters in their username to build successfully when using exec (#6223)
Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1264667

Context
Users with profile names that contain parens in them will never be able to succesfully build a program that calls Exec with any command. This is because exec does the following:

Generate a cmd file in the users temp directory
Calls cmd /Q /D /C <path-to-temp-file>
The problem with this is that running cmd /C "some command" does not work if the command has any parens in it. It needs to be escaped like so: ^( and ^).

Changes Made
When user sets EscapeSpecialCharacters in the exec task (boolean parameter), we escape characters that need to be escaped when calling cmd /c. We preserve the original functionality of always removing spaces and escaping '^'
Added under changewave 16.10
Added documentation for it.