Fix TasksCanAddRecursiveDirBuiltInMetadata()
Don't use Program Files\Reference Assemblies as that might not exist for 64-bit.

Pass ITestOutputLogger to MockLogger to get the full log in case of failure.

Search for a log substring that will only be there in case of success. Even successful logs can contain [] when the file is at the root of the directory so RecursiveDir will be empty.