Allow turning detailed summary off
Change the DetailedSummary command-line switch from parameterless to parameterized, allowing the user to have control and explicitly set it to false if needed, overriding the default of true when the verbosity is diagnostic.

This is backwards compatible with the previous behavior, so if /ds is specified, it is equivalent to turning it on.

Fixes https://github.com/dotnet/msbuild/issues/4409