Add support for embedding arbitrary files into binlog
It would be very useful to embed more than just the project files into the binlog. For example, project.assets.json is a common request: https://github.com/dotnet/msbuild/issues/3529. global.json or nuget.config would be other examples. Some users may want to embed all *.cs files as well.

Add a special EmbedInBinlog item type recognized by BuildEventArgsWriter. If after evaluation the items contain EmbedInBinlog, all such items will be logged if the evaluated include points to a full file path that exists on disk. A given file can be mentioned more than once, but will only be embedded once. If a file doesn't exist or is empty, it will not be embedded. If the item is added from a target (after evaluation is done), it will not be embedded.

Checking for EmbedInBinlog is done in BuildEventArgsWriter to reuse the complicated and performance-optimized logic of iterating over all items.