Change default DebugType from 'pdbonly' to 'full'
Per [this documentation](https://github.com/dotnet/roslyn/blob/main/docs/compilers/CSharp/CommandLine.md):

> `/debug:pdbonly` &mdash; Same as `/debug:full`. For backward compatibility.

In the new Project Properties UI we want to hide "pdbonly" from the list of options presented to the user. It would be better if the web SDK did not use it as a default.