Make GenerateResource look up resgen.exe in 10.0 SDK (#6336)
* Make GenerateResource look up resgen.exe in 10.0 SDK

GenerateResource tests were failing when only VS 2019 and 10.0 Windows SDK is installed. I suspect the task itself would also fail if run with ExecuteAsTool.

The problem is that we started look up with .NET 4.6.1 and lower. Start lookup with .NET 4.8 instead and fallback down the chain. This should still find all earlier SDKs too.

Introduce latest versions of Visual Studio too.

* Rollback some changes to VersionLatest.

* Set the SdkToolsPath to framework 4.8

This actually fixes the unit-tests on a machine with only SDK 10.0 installed.