Snapshot Items of TaskParameters from another AppDomain
Fixes https://github.com/dotnet/msbuild/issues/6379

We have a situation where we are in a worker node and a task runs in a separate AppDomain and logs a TaskParameterEventArgs. Since logging is asynchronous there's a risk that by the time the node packet translator accesses the TaskParameterEventArgs.Items the AppDomain is already unloaded and we crash when trying to enumerate item metadata.

Detect that we're in another AppDomain and eagerly take a snapshot of task items with all metadata.