Don't send yield-return iterators across AppDomain boundaries
To access the metadata of items from another AppDomain we can't use yield return because the generated iterator class is not marked as [Serializable]. Provide eager codepaths for cross-AppDomain scenarios that return simple arrays.

We can't always return simple arrays because in the general case we want to stream metadata directly to consumers without allocating a collection to store the results.