Log location of intrinsic tasks
Log the Line and Column of add or remove items in ItemGroups inside targets. Previously we could only recover the file from the parent target, but not exactly which line in the file corresponded to the add/remove.

Now we can exactly pinpoint which add/remove is executing. No need to log the file because we can recover this information from the parent target. This saves space.