Support multiple rights vs same left scenario with a simple call to ApiComparer (#17295)
* Support multiple rights vs same left scenario with a simple call to
ApiComparer.

Also, change how the Rules are loaded and invoked, by following a
similar model as Roslyn, where an Initialize method is invoked and we
pass in a context. Then Rules can register to context callbacks, that
way the Rule just register to whatever callbacks it is important to run
on based on the Mappers it runs for.