Add solution-validation targets as hook points
Solution metaprojects are assembled in a confusing way that impacts what
targets can be hooked into via BeforeTargets/AfterTargets. Prior to this
change one couldn't hook before the build would attempt to traverse the
list of projects in the solution.

Extend the "list of targets we know we'll create shortly" list to
include the things that are in the `InitialTargets` list and
`GetSolutionConfigurationContents`.

Fixes #6452.