Schedule proxy builds to inproc node (#6386)
Context
Proxy builds are what project caches issue on cache hits. They are a cheap version of the expensive targets that were avoided by the cache. They need to produce the same properties and items the expensive target produced, but with none of the CPU / IO expensive stuff.

The proxy builds are super cheap because they only return properties / items. It is not worth scheduling them to out of proc nodes because:

IPC overhead
when they get scheduled to out of proc nodes they get re-evaluated. This is wasted computation because proxy builds are guaranteed to get evaluated on the scheduler node (where the inproc node resides)
Scheduling proxy builds to the inproc node makes a project cache build with full cache hits 16% faster.

Changes Made
Duplicated what the scheduler does to confine traversal projects to the inproc node, since those are also cheap.