Don't move Properties and Items to ProjectEvaluationFinished if legacy loggers present
Switch from the "use the new logic if any logger is present that supports it" to the more conservative "use the old logic if any logger doesn't support the new logic".

There are legacy loggers such as the Azure DevOps logger that crash if ProjectStartedEventArgs.Properties is null.

Both console loggers also need more work to properly support the new logic.

Effectively the new logic will now only take place when the binary logger is the only logger.