Issue templates apply needs-triage (#6557)
Updating untriaged to needs-triage was done for consistency (needs-attention/design/more-info), ease of search/use (no spaces, all lowercase). Not to mention `untriaged` eventually meant some combination of "find a bucket this fits in" and "this needs attention".

However updating the label requires an update to our bug templates that apply those labels.