Get DependencyModel from the LKG SDK (#6548)
Normally DependencyModel is resolvable in dotnet build scenarios because
it is in the SDK folder next to MSBuild. However, in our
bootstrap MSBuild, that wasn't the case. Grab that file from the SDK
that we're using to build and drop it next to the SDK task assemblies (not 
next to MSBuild, because of #6558).